+++ common/src/main/java/it/silleellie/dndsync/shared/WearSignal.java
@@ -1,4 +1,4 @@
-package it.silleellie.dndsync;
+package it.silleellie.dndsync.shared;
 
 import java.io.Serializable;
